+++ mkdocs.yml
@@ -1,4 +1,4 @@
-site_name: HUMPY - Mixed Graph Estimation
+site_name: HUMLPY - Mixed Graph Learning
 
 repo_url: https://github.com/konstantingoe/mixed-gm
 site_url: https://konstantingoe.github.io/mixed-gm/
