+++ requirements.txt
@@ -5,15 +5,15 @@ cycler==0.12.1            # via matplotlib
 fonttools==4.61.1         # via matplotlib
 joblib==1.5.3             # via scikit-learn
 kiwisolver==1.4.9         # via matplotlib
-matplotlib==3.10.8        # via humpy (pyproject.toml)
-networkx==3.6.1           # via humpy (pyproject.toml)
-numpy==2.4.2              # via contourpy, matplotlib, pandas, scikit-learn, scipy, humpy (pyproject.toml)
+matplotlib==3.10.8        # via humlpy (pyproject.toml)
+networkx==3.6.1           # via humlpy (pyproject.toml)
+numpy==2.4.2              # via contourpy, matplotlib, pandas, scikit-learn, scipy, humlpy (pyproject.toml)
 packaging==26.0           # via matplotlib
-pandas==3.0.1             # via humpy (pyproject.toml)
+pandas==3.0.1             # via humlpy (pyproject.toml)
 pillow==12.1.1            # via matplotlib
 pyparsing==3.3.2          # via matplotlib
 python-dateutil==2.9.0.post0  # via matplotlib, pandas
-scikit-learn==1.8.0       # via humpy (pyproject.toml)
-scipy==1.17.1             # via scikit-learn, humpy (pyproject.toml)
+scikit-learn==1.8.0       # via humlpy (pyproject.toml)
+scipy==1.17.1             # via scikit-learn, humlpy (pyproject.toml)
 six==1.17.0               # via python-dateutil
 threadpoolctl==3.6.0      # via scikit-learn
